feat - add cleanup and write image to jpg
