+++ backend/profile-service/src/main/java/org/maxq/profileservice/config/image/ImageServiceConfig.java
@@ -6,6 +6,9 @@ import org.apache.commons.imaging.formats.jpeg.xmp.JpegXmpRewriter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import javax.imageio.ImageIO;
+import javax.imageio.ImageWriter;
+
 @Configuration
 public class ImageServiceConfig {
 
@@ -22,5 +25,10 @@ public class ImageServiceConfig {
   @Bean
   public JpegIptcRewriter jpegIptcRewriter() {
     return new JpegIptcRewriter();
+  }
+
+  @Bean
+  public ImageWriter jpegImageWriter() {
+    return ImageIO.getImageWritersByFormatName("jpeg").next();
   }
 }
