+++ backend/profile-service/src/main/java/org/maxq/profileservice/domain/ImageMetadata.java
@@ -2,9 +2,11 @@ package org.maxq.profileservice.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.ToString;
 
 @Getter
 @RequiredArgsConstructor
+@ToString
 public class ImageMetadata {
 
   private final ImageSize size;
