+++ backend/profile-service/src/test/java/org/maxq/profileservice/service/message/handler/ProfileImageUploadHandlerTest.java
@@ -7,6 +7,7 @@ import org.maxq.profileservice.domain.InMemoryFile;
 import org.maxq.profileservice.domain.dto.ImageDto;
 import org.maxq.profileservice.mapper.InMemoryFileMapper;
 import org.maxq.profileservice.service.image.ApacheImageService;
+import org.maxq.profileservice.service.image.processor.ApacheImageProcessor;
 import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -28,6 +29,8 @@ class ProfileImageUploadHandlerTest {
   private InMemoryFileMapper inMemoryFileMapper;
   @MockitoBean
   private ApacheImageService imageService;
+  @MockitoBean
+  private ApacheImageProcessor imageProcessor;
 
   @Mock
   private BufferedImage mockImage;
@@ -48,15 +51,17 @@ class ProfileImageUploadHandlerTest {
         imageDto.getData()
     );
     when(inMemoryFileMapper.mapToInMemoryFile(imageDto)).thenReturn(file);
-    when(imageService.stripMetadata(file)).thenReturn(file);
-    when(imageService.resizeImage(file)).thenReturn(mockImage);
+    when(imageProcessor.stripMetadata(file)).thenReturn(file);
+    when(imageProcessor.resizeImage(file)).thenReturn(mockImage);
+    when(imageProcessor.cleanImage(mockImage)).thenReturn(mockImage);
+    when(imageService.writeToJpeg(mockImage)).thenReturn(file);
 
     // When
     handler.handleMessage(imageDto);
 
     // Then
-    verify(imageService, times(1)).stripMetadata(file);
-    verify(imageService, times(1)).resizeImage(file);
+    verify(imageProcessor, times(1)).stripMetadata(file);
+    verify(imageProcessor, times(1)).resizeImage(file);
   }
 
   @Test
@@ -69,14 +74,14 @@ class ProfileImageUploadHandlerTest {
         imageDto.getData()
     );
     when(inMemoryFileMapper.mapToInMemoryFile(imageDto)).thenReturn(file);
-    doThrow(IOException.class).when(imageService).stripMetadata(file);
+    doThrow(IOException.class).when(imageProcessor).stripMetadata(file);
 
     // When + Then
     Executable executable = () -> handler.handleMessage(imageDto);
 
     // Then
     assertDoesNotThrow(executable);
-    verify(imageService, times(1)).stripMetadata(file);
+    verify(imageProcessor, times(1)).stripMetadata(file);
   }
 
   @Test
@@ -89,14 +94,38 @@ class ProfileImageUploadHandlerTest {
         imageDto.getData()
     );
     when(inMemoryFileMapper.mapToInMemoryFile(imageDto)).thenReturn(file);
-    when(imageService.stripMetadata(file)).thenReturn(file);
-    doThrow(IOException.class).when(imageService).resizeImage(file);
+    when(imageProcessor.stripMetadata(file)).thenReturn(file);
+    doThrow(IOException.class).when(imageProcessor).resizeImage(file);
 
     // When + Then
     Executable executable = () -> handler.handleMessage(imageDto);
 
     // Then
     assertDoesNotThrow(executable);
-    verify(imageService, times(1)).stripMetadata(file);
+    verify(imageProcessor, times(1)).stripMetadata(file);
+  }
+
+  @Test
+  void shouldHandleMessage_When_WritingFailed() throws IOException {
+    // Given
+    ImageDto imageDto = new ImageDto("test@test.com", "test.jpeg", "image/jpeg", "test-data".getBytes());
+    InMemoryFile file = new InMemoryFile(
+        imageDto.getContentType(),
+        imageDto.getName(),
+        imageDto.getData()
+    );
+    when(inMemoryFileMapper.mapToInMemoryFile(imageDto)).thenReturn(file);
+    when(imageProcessor.stripMetadata(file)).thenReturn(file);
+    when(imageProcessor.resizeImage(file)).thenReturn(mockImage);
+    when(imageProcessor.cleanImage(mockImage)).thenReturn(mockImage);
+    doThrow(IOException.class).when(imageService).writeToJpeg(mockImage);
+
+
+    // When
+    Executable executable = () -> handler.handleMessage(imageDto);
+
+    // Then
+    assertDoesNotThrow(executable);
+    verify(imageService, times(1)).writeToJpeg(mockImage);
   }
 }
